nixutils: move NarInfo representation into the crate
The NarInfo narinfo-file representation — the struct, its parse/serialize
logic, the Compression enum, and the signature-trust predicates — lived in
the ogygia-irisd binary crate, even though it is a reusable Nix data format
that belongs with the other store types and their tests in ogygia-nixutils.

This moves NarInfo into ogygia-nixutils and adds the strum dependency there.
Parsing and serialization are now exposed as the standard FromStr and Display
traits (rather than inherent parse/serialize methods), matching the sibling
StoreHash and NarHash types in the same crate.

narinfo_from_path_info stays in ogygia-irisd as a free function: it is not a
property of the format but irisd's serving policy, choosing zstd compression
and a NAR URL with the NarHash embedded, and leaving placeholder FileHash /
FileSize for the caller to fill in after generating the NAR. All NarInfo
fields are public, so it builds the struct directly.

Test plan:
- The narinfo format tests (parse, serialize roundtrip via Display, is_trusted
  / has_trusted_signature) moved with the type and pass under ogygia-nixutils.
- The narinfo_from_path_info test moved with the function and passes under
  ogygia-irisd.
- nix flake check passes.
