+++ src/ogygia-nixutils/Cargo.toml
@@ -12,6 +12,7 @@ base64 = { workspace = true }
 hex = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
+strum = { workspace = true }
 
 [dev-dependencies]
 tempfile = { workspace = true }
